requirements.txt
streamlit==1.26.0
pandas==2.0.3
google-generativeai==0.3.0